Move pre processing code to a header
