+++ tools/replay/android_main.cpp
@@ -37,6 +37,7 @@
 #include "util/logging.h"
 #include "util/platform.h"
 #include "parse_dump_resources_cli.h"
+#include "replay_pre_processing.h"
 
 #include <android_native_app_glue.h>
 #include <android/log.h>
@@ -61,73 +62,6 @@ int32_t     ProcessInputEvent(struct android_app* app, AInputEvent* event);
 void        DestroyActivity(struct android_app* app);
 
 static std::unique_ptr<gfxrecon::decode::FileProcessor> file_processor;
-
-struct ApiReplayOptions
-{
-    gfxrecon::decode::VulkanReplayOptions* vk_replay_options{ nullptr };
-};
-
-struct ApiReplayConsumer
-{
-    gfxrecon::decode::VulkanReplayConsumer* vk_replay_consumer{ nullptr };
-};
-
-bool IsRunPreProcessConsumer(ApiReplayOptions& replay_options)
-{
-    GFXRECON_ASSERT(replay_options.vk_replay_options != nullptr);
-
-    if (replay_options.vk_replay_options->enable_dump_resources)
-    {
-        return true;
-    }
-    return false;
-}
-
-void RunPreProcessConsumer(const std::string& input_filename,
-                           ApiReplayOptions&  replay_options,
-                           ApiReplayConsumer& replay_consumer)
-{
-    GFXRECON_ASSERT(replay_options.vk_replay_options != nullptr && replay_consumer.vk_replay_consumer != nullptr);
-
-    gfxrecon::decode::FileProcessor file_processor;
-    if (file_processor.Initialize(input_filename))
-    {
-        gfxrecon::decode::VulkanPreProcessConsumer vk_pre_process_consumer;
-
-        if (replay_options.vk_replay_options->using_dump_resources_target)
-        {
-            vk_pre_process_consumer.EnableDumpResources(replay_options.vk_replay_options->dump_resources_target);
-        }
-
-        gfxrecon::decode::VulkanDecoder vk_decoder;
-        vk_decoder.AddConsumer(&vk_pre_process_consumer);
-        file_processor.AddDecoder(&vk_decoder);
-
-        file_processor.ProcessAllFrames();
-
-        replay_options.vk_replay_options->enable_vulkan = vk_pre_process_consumer.WasVulkanAPIDetected();
-
-        if (replay_options.vk_replay_options->enable_vulkan)
-        {
-            if (replay_options.vk_replay_options->using_dump_resources_target)
-            {
-                replay_options.vk_replay_options->dump_resources_block_indices =
-                    vk_pre_process_consumer.GetDumpResourcesBlockIndices();
-            }
-
-            if (replay_options.vk_replay_options->enable_dump_resources)
-            {
-                // Process --dump-resources block indices arg.
-                if (!gfxrecon::parse_dump_resources::parse_dump_resources_arg(*replay_options.vk_replay_options))
-                {
-                    GFXRECON_LOG_FATAL("There was an error while parsing dump resources indices. Terminating.");
-                    exit(0);
-                }
-                replay_consumer.vk_replay_consumer->InitializeReplayDumpResources();
-            }
-        }
-    }
-}
 
 extern "C"
 {
